Edit InputHandler.java
